Refactor dataset loader op to avoid duplicate definitions
