+++ dagster_app/dagster_app/ops/__init__.py
@@ -1,0 +1,5 @@
+"""Reusable Dagster ops for the dagster_app service."""
+
+from .dataset import load_dataset_op
+
+__all__ = ["load_dataset_op"]
