--- dagster_app/data/home_credit/application_train.csv
@@ -1,4 +1,0 @@
-SK_ID_CURR,TARGET,AMT_INCOME_TOTAL,AMT_CREDIT
-100001,1,202500.0,406597.5
-100002,0,270000.0,1293502.5
-100003,0,67500.0,135000.0
