--- dagster_app/data/home_credit/bureau.csv
@@ -1,4 +1,0 @@
-SK_ID_CURR,SK_ID_BUREAU,CREDIT_ACTIVE,AMT_CREDIT_SUM
-100001,571982,Active,3000.0
-100002,571983,Closed,5000.0
-100004,571984,Bad debt,1000.0
